[django] split our huge GraphQL script into 3 parts : "query", "schema" and "utils"
There is still room for improvement in the way we define and implement our GraphQL queries, but it's already a bit better like that :-)
