+++ server/api/django/apps/api_public/graphql_schema.py
@@ -2,12 +2,12 @@ import graphene
 from django.conf import settings
 from graphene_django.debug import DjangoDebug
 
-import api_public.graphql.library as schema_library
+from api_public.graphql.library.query import Query as LibraryQuery
 
 
 # This class will inherit from multiple Queries
 # as we begin to add more apps to our project
-class Query(schema_library.Query, graphene.ObjectType):
+class Query(LibraryQuery, graphene.ObjectType):
     debug = graphene.Field(DjangoDebug, name='__debug') if settings.DEBUG else None
